detail con json-server
