+++ src/components/ListNft/index.jsx
@@ -1,8 +1,26 @@
 import './styles.scss';
+import { useState, useEffect } from 'react';
 import NftCard from '../NftCard';
 
+import { getAllProducts } from '../../hook/products';
+
 // eslint-disable-next-line react/prop-types
-const ListNft = ({ products = [] }) => {
+const ListNft = () => {
+  const [products, setProducts] = useState([]);
+
+  useEffect(() => {
+    const getData = async () => {
+      try {
+        const data = await getAllProducts();
+        setProducts(data);
+      } catch (error) {
+        console.log(error);
+      }
+    };
+
+    getData();
+  }, []);
+
   return (
     <div className="products__container">
       {
