+++ src/hook/products.jsx
@@ -1,11 +1,13 @@
+const API = 'http://localhost:3004';
+
 export async function getAllProducts() {
-  const res = await fetch('192.168.128.7:3000/api/products');
+  const res = await fetch(`${API}/products`);
   const result = await res.json();
   return result;
 }
 
 export async function getProduct(id) {
-  const res = await fetch(`192.168.128.7:3000/api/products/${id}`);
+  const res = await fetch(`${API}/products/${id}`);
   const result = await res.json();
   return result;
 }
